Rename 'AblyRealtimeSpecs' files to match type name
Tidy type layout along BCL FDG lines.
